+++ evm/interfaces/dexs/IThenaV3Voter.sol
@@ -7,16 +7,16 @@ interface IThenaV3Voter {
     function isGauge(address _gauge) external view returns (bool);
     function poolForGauge(address _gauge) external view returns (address);
     function factory() external view returns (address);
-    function minter() external view returns(address);
+    function minter() external view returns (address);
     function isWhitelisted(address token) external view returns (bool);
-    function notifyRewardAmount(uint amount) external;
+    function notifyRewardAmount(uint256 amount) external;
     function distributeAll() external;
     function distributeFees(address[] memory _gauges) external;
     function internal_bribes(address _gauge) external view returns (address);
     function external_bribes(address _gauge) external view returns (address);
-    function usedWeights(uint id) external view returns(uint);
-    function lastVoted(uint id) external view returns(uint);
-    function poolVote(uint id, uint _index) external view returns(address _pair);
-    function votes(uint id, address _pool) external view returns(uint votes);
-    function poolVoteLength(uint tokenId) external view returns(uint);
+    function usedWeights(uint256 id) external view returns (uint256);
+    function lastVoted(uint256 id) external view returns (uint256);
+    function poolVote(uint256 id, uint256 _index) external view returns (address _pair);
+    function votes(uint256 id, address _pool) external view returns (uint256 votes);
+    function poolVoteLength(uint256 tokenId) external view returns (uint256);
 }
